+++ src/Pages/Products/ProductsCard.jsx
@@ -1,23 +1,9 @@
-import React, { useContext, useState } from 'react';
+import React, { useContext } from 'react';
 import { Link } from 'react-router-dom';
 import { CartContext } from '../../App';
 
 const ProductCard = ({ product }) => {
-  const [cartItems, setCartItems] = useContext(CartContext);
-  console.log(cartItems)
-  const handleAddToCart = (clickedItem) => {
-    setCartItems((prev) => {
-      const isItemInCart = prev.find((item) => item.productId === clickedItem.productId);
-      if (isItemInCart) {
-        return prev.map((item) =>
-          item.productId === clickedItem.productId
-            ? { ...item, amount: item.amount + 1 }
-            : item
-        );
-      }
-      return [...prev, { ...clickedItem, amount: 1 }];
-    });
-  };
+  const { handleAddToCart } = useContext(CartContext);
   return (
     <div className="flex flex-col justify-center items-center">
       <div className="flex justify-center w-[660px] h-[2px] bg-primary-default my-6" />
@@ -33,7 +19,7 @@ const ProductCard = ({ product }) => {
             <input className='px-4 text-background-main select-none' id="quantity" type="number" min="1" max="5" placeholder='01' />
           </div>
           <div className='flex flex-col mt-2'>
-            <button onClick={handleAddToCart(product)} className='px-4 bg-white text-background-main font-semibold hover:opacity-80 rounded'>Add to cart</button>
+            <button onClick={() => handleAddToCart(product)} className='px-4 bg-white text-background-main font-semibold hover:opacity-80 rounded'>Add to cart</button>
           </div>
         </div>
       </div>
